Moved src/Table/insert_ascii_in_row.cxx to Utils directory
